refactor: 重构楼层评论数

+++ upupor-service/src/main/java/com/upupor/service/outer/req/AddCommentReq.java
@@ -62,5 +62,5 @@ public class AddCommentReq {
     /**
      * 被回复的楼层
      */
-    private Integer beFloorNum;
+    private Long beFloorNum;
 }
